adicionadas novas animations para a layer, possivel passar float pro tamanho da imagem, factory de popups para mostrar um poup apos o outro sem encavalar popup

+++ JewFeatures/Classes/CommonKit/Extensions/UIImage+JEWExtension.swift
@@ -53,8 +53,8 @@ public extension UIImage {
     }
     
     
-    func compressTo(_ expectedSizeInMb:Int) -> Data? {
-        let sizeInBytes = expectedSizeInMb * 1024 * 1024
+    func compressTo(_ expectedSizeInMb:Float) -> Data? {
+        let sizeInBytes = Int(expectedSizeInMb * 1024 * 1024)
         var needCompress:Bool = true
         var imgData:Data?
         var compressingValue:CGFloat = 1.0
